pyproject.toml
# SPDX-FileCopyrightText: 2025 Battelle Memorial Institute
# SPDX-License-Identifier: BSD-2-Clause
[project]
name = "hybridlane"
version = "0.6.1a1"
authors = [
    {name = "Jim Furches", email = "james.furches@pnnl.gov"}
]
description = "A frontend library for hybrid CV-DV computation based on Pennylane."
readme = "README.md"
license = "BSD-2-Clause"
license-files = ["LICENSE.txt", "DISCLAIMER.txt"]
classifiers = [
    "Development Status :: 3 - Alpha",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.11"
dependencies = [
    "pennylane>=0.44.0",
    "pydantic>=2.9.2",
]

[project.urls]
Homepage = "https://github.com/pnnl/hybridlane"
Documentation = "https://pnnl.github.io/hybridlane/"
Repository = "https://github.com/pnnl/hybridlane.git"
Issues = "https://github.com/pnnl/hybridlane/issues"

[project.optional-dependencies]
bq = [
    "bosonic-qiskit>=15.0",
    "sparse>=0.17.0",
]
all = [
    "hybridlane[bq,qscout]"
]
qscout = [
    "jaqalpaq>=1.3.0",
    "python-constraint2>=2.4.0",
    "qscout-gatemodels>=1.3.0",
]

[project.entry-points."pennylane.plugins"]
"bosonicqiskit.hybrid" = "hybridlane.devices:BosonicQiskitDevice"
"sandiaqscout.hybrid" = "hybridlane.devices:QscoutIonTrap"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "openqasm3[parser]>=1.0.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
    "sybil[pytest]>=9.3.0",
]
docs = [
    "furo>=2024.8.6",
    "sphinx>=8.1.3",
    "sphinx-autoapi>=3.6.0",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-copybutton>=0.5.2",
    "sphinx-math-dollar>=1.2.1",
    "sphinxcontrib-bibtex>=2.6.5",
]
